move VTV to include file
