+++ addons/MH60M/config/cfgVehicles.hpp
@@ -14,7 +14,7 @@ class CfgVehicles {
     class Helicopter_Base_H;
     class Heli_Transport_01_base_F: Helicopter_Base_H {
         class Turrets;
-    }; // Helicopter_Base_H
+    }; // Heli_Transport_01_base_F
     class vtx_H60_base: Heli_Transport_01_base_F {
         class ViewPilot;
         class vtx_templateFLIR;
@@ -25,7 +25,7 @@ class CfgVehicles {
             class MainTurret;
             class CopilotTurret;
         };
-    }; // Heli_Transport_01_base_F
+    }; // vtx_H60_base
     class vtx_MH60M: vtx_H60_base {
         class pilotCamera: vtx_templateFLIR {};
         driverWeaponsInfoType = "Rsc_MELB_Turret_UnitInfo";
@@ -103,21 +103,7 @@ class CfgVehicles {
             #include "\z\vtx\addons\UH60\config\turrets\cargoTurrets.hpp"
         };
         hiddenSelectionsTextures[] = {"","","","","","","","","","","","","","","","","z\vtx\addons\MH60M\data\main_co.paa","z\vtx\addons\MH60M\Data\Misc_co.paa","z\vtx\addons\MH60M\data\tail_co.paa","","","","","","z\vtx\addons\UH60\data\FuelProbe\Fuel_probe_co.paa"};
-        class VehicleTransport {
-        	class Carrier {
-        		cargoBayDimensions[]		= {"VTV_limit_1", "VTV_limit_2"};	// Memory points in model defining cargo space
-        		disableHeightLimit			= 0;								// If set to 1 disable height limit of transported vehicles
-        		maxLoadMass					= 4100;							// Maximum cargo weight (in Kg) which the vehicle can transport
-        		cargoAlignment[]			= {"front", "center"};				// Array of 2 elements defining alignment of vehicles in cargo space. Possible values are left, right, center, front, back. Order is important.
-        		cargoSpacing[]				= {0.3, 0.3, 0};						// Offset from X,Y,Z axes (in metres)
-        		exits[]						= {"pos Cargo L", "pos Cargo R"};		// Memory points in model defining loading ramps, could have multiple
-        		unloadingInterval			= 2;								// Time between unloading vehicles (in seconds)
-        		loadingDistance				= 5;								// Maximal distance for loading in exit point (in meters).
-        		loadingAngle				= 60;								// Maximal sector where cargo vehicle must be to for loading (in degrees).
-        		parachuteClassDefault       = B_Parachute_02_F;					// Type of parachute used when dropped in air. Can be overridden by parachuteClass in Cargo.
-        		parachuteHeightLimitDefault = 50;								// Minimal height above terrain when parachute is used. Can be overriden by parachuteHeightLimit in Cargo.
-        	};
-        };
+        #include "\z\vtx\addons\UH60\config\vehicleTransport.hpp"
     }; // vtx_MH60M
     class vtx_MH60M_DAP : vtx_MH60M {
         class pilotCamera: vtx_templateFLIR {};
